requirementst.txt
asciichartpy==1.5.25
asttokens==2.2.1
backcall==0.2.0
bashplotlib==0.6.5
bcrypt==4.0.1
black==23.9.1
blessings==1.7
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
crayons==0.4.0
cryptography==41.0.1
cycler==0.12.0
Cython==3.0.0
darkdetect==0.7.1
decorator==5.1.1
diff-match-patch==20230430
docutils==0.20.1
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.2.2
fonttools==4.43.0
future==0.18.3
graphml2svg==0.0.4
greenlet==2.0.2
idle==1.0.4
idna==3.4
igraph==0.10.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
keyring==24.2.0
kiwisolver==1.4.5
ldap3==2.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
n2g==0.3.3
netmiko==4.2.0
nh3==0.2.14
ntc-templates==3.5.0
numpy==1.25.2
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pandasql==0.7.3
paramiko==3.2.0
parsec==3.15
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.1
pkginfo==1.9.6
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psycopg2==2.9.7
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.15.1
PyNaCl==1.5.0
pynetbox==7.2.0
pyparsing==3.1.1
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
PyQt6-WebEngine==6.5.0
PyQt6-WebEngine-Qt6==6.5.1
pyqtdarktheme==2.1.0
pyserial==3.5
pyte==0.8.1
python-dateutil==2.8.2
python-igraph==0.10.6
pytz==2023.3
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0.1
qt-material==2.14
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
scp==0.14.5
six==1.16.0
SQLAlchemy==2.0.19
sqlalchemy-orm==1.2.10
SQLAlchemy-serializer==1.4.1
stack-data==0.6.2
svgwrite==1.4.3
textfsm==1.1.3
texttable==1.6.7
tftpy==0.8.2
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
ttp==0.9.5
ttp-templates==0.3.5
twine==4.0.2
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unqlite==0.9.6
urllib3==2.0.4
wcwidth==0.2.6
webssh==1.6.2
Werkzeug==3.0.0
windows-curses==2.3.1
zipp==3.17.0
